requirements.txt
# 宏观经济分析投资决策系统依赖
# 基础依赖
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0

# 数据库连接
mysql-connector-python>=8.0.26
sqlalchemy>=1.4.0

# 配置和日志
pyyaml>=6.0
python-dotenv>=0.19.0

# 数据获取
requests>=2.26.0
feedparser>=6.0.0
newsapi-python>=0.2.6
beautifulsoup4>=4.10.0
lxml>=4.6.0

# 文本分析
textblob>=0.15.3
nltk>=3.6.0
scikit-learn>=1.0.0

# 数据可视化
plotly>=5.3.0
matplotlib-venn>=0.11.6

# 报告生成
jinja2>=3.0.0
weasyprint>=53.0
fpdf2>=2.5.0
openpyxl>=3.0.0

# 测试
pytest>=6.2.0
pytest-cov>=2.12.0

# 开发工具
black>=21.5b2
flake8>=3.9.0
isort>=5.9.0
